requirements-prod.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
torch==1.13.1
torchvision==0.14.1
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
tqdm==4.66.1
openai==0.28.1
python-dotenv==1.0.0
gunicorn==21.2.0
